# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StartGameRequest {
  /** Optional player name */
  playerName?: string;
}

export interface ActionRequest {
  /** The voice command or action string (e.g. "attack goblin", "move north") */
  command: string;
}

export interface ParsedCommand {
  /** The parsed action type (e.g. ATTACK, MOVE, DEFEND, CAST_SPELL, LOOK) */
  action: string;
  /** The parsed target if any */
  target?: string;
  /** Direction for movement commands */
  direction?: string;
  /** The original raw command string */
  raw: string;
}

export interface PlayerState {
  id: string;
  name: string;
  hp: number;
  maxHp: number;
  mp: number;
  maxMp: number;
  level: number;
  xp: number;
  xpToNextLevel: number;
  attack: number;
  defense: number;
  abilities: string[];
  inventory: string[];
}

export interface EnemyState {
  id: string;
  name: string;
  hp: number;
  maxHp: number;
  attack: number;
  defense: number;
  xpReward: number;
  isDefeated: boolean;
}

export type RoomStateExits = { [key: string]: string };

export interface RoomState {
  id: string;
  name: string;
  description: string;
  exits: RoomStateExits;
  enemies: EnemyState[];
  items: string[];
  isExplored: boolean;
}

export interface GameStateResponse {
  sessionId: string;
  /** IDLE, IN_COMBAT, EXPLORING, GAME_OVER, VICTORY */
  gameStatus: string;
  player: PlayerState;
  currentRoom: RoomState;
  /** Narrative messages describing what happened */
  logs: string[];
  /** The command that was parsed (only present on action responses) */
  parsedCommand?: ParsedCommand;
  turnCount: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}