# lib/api-zod/src/generated/types/actionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActionRequest {
  /** The voice command or action string (e.g. "attack goblin", "move north") */
  command: string;
}