# lib/api-zod/src/generated/types/enemyState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnemyState {
  id: string;
  name: string;
  hp: number;
  maxHp: number;
  attack: number;
  defense: number;
  xpReward: number;
  isDefeated: boolean;
}