# lib/api-zod/src/generated/types/gameStateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedCommand } from "./parsedCommand";
import type { PlayerState } from "./playerState";
import type { RoomState } from "./roomState";

export interface GameStateResponse {
  sessionId: string;
  /** IDLE, IN_COMBAT, EXPLORING, GAME_OVER, VICTORY */
  gameStatus: string;
  player: PlayerState;
  currentRoom: RoomState;
  /** Narrative messages describing what happened */
  logs: string[];
  /** The command that was parsed (only present on action responses) */
  parsedCommand?: ParsedCommand;
  turnCount: number;
}