# lib/api-zod/src/generated/types/parsedCommand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParsedCommand {
  /** The parsed action type (e.g. ATTACK, MOVE, DEFEND, CAST_SPELL, LOOK) */
  action: string;
  /** The parsed target if any */
  target?: string;
  /** Direction for movement commands */
  direction?: string;
  /** The original raw command string */
  raw: string;
}