# lib/api-zod/src/generated/types/playerState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerState {
  id: string;
  name: string;
  hp: number;
  maxHp: number;
  mp: number;
  maxMp: number;
  level: number;
  xp: number;
  xpToNextLevel: number;
  attack: number;
  defense: number;
  abilities: string[];
  inventory: string[];
}