# lib/api-zod/src/generated/types/roomState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnemyState } from "./enemyState";
import type { RoomStateExits } from "./roomStateExits";

export interface RoomState {
  id: string;
  name: string;
  description: string;
  exits: RoomStateExits;
  enemies: EnemyState[];
  items: string[];
  isExplored: boolean;
}