# lib/api-zod/src/generated/types/startGameRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dora Dungeons API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StartGameRequest {
  /** Optional player name */
  playerName?: string;
}